+++ server/src/api-v1/asset/asset.module.ts
@@ -3,8 +3,6 @@ import { AssetService } from './asset.service';
 import { AssetController } from './asset.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AssetEntity } from './entities/asset.entity';
-import { ImageOptimizeModule } from '../../modules/image-optimize/image-optimize.module';
-import { AssetOptimizeService } from '../../modules/image-optimize/image-optimize.service';
 import { BullModule } from '@nestjs/bull';
 import { BackgroundTaskModule } from '../../modules/background-task/background-task.module';
 import { BackgroundTaskService } from '../../modules/background-task/background-task.service';
@@ -39,11 +37,10 @@ import { CommunicationModule } from '../communication/communication.module';
       },
     }),
     TypeOrmModule.forFeature([AssetEntity]),
-    ImageOptimizeModule,
     BackgroundTaskModule,
   ],
   controllers: [AssetController],
-  providers: [AssetService, AssetOptimizeService, BackgroundTaskService],
+  providers: [AssetService, BackgroundTaskService],
   exports: [],
 })
 export class AssetModule { }
