+++ server/src/modules/schedule-tasks/image-conversion.service.ts
@@ -13,40 +13,40 @@ export class ImageConversionService {
     private assetRepository: Repository<AssetEntity>
   ) { }
 
-  @Cron(CronExpression.EVERY_5_MINUTES
-    , {
-      name: 'webp-conversion'
-    })
-  async webpConversion() {
-    Logger.log('Starting Webp Conversion Tasks', 'ImageConversionService')
+  // @Cron(CronExpression.EVERY_5_MINUTES
+  //   , {
+  //     name: 'webp-conversion'
+  //   })
+  // async webpConversion() {
+  //   Logger.log('Starting Webp Conversion Tasks', 'ImageConversionService')
 
-    const assets = await this.assetRepository.find({
-      where: {
-        webpPath: ''
-      },
-      take: 500
-    });
-
-
-    if (assets.length == 0) {
-      Logger.log('All assets has webp file - aborting task', 'ImageConversionService')
-      return;
-    }
+  //   const assets = await this.assetRepository.find({
+  //     where: {
+  //       webpPath: ''
+  //     },
+  //     take: 500
+  //   });
 
 
-    for (const asset of assets) {
-      const resizePath = asset.resizePath;
-      if (resizePath != '') {
-        const webpPath = resizePath.replace('jpeg', 'webp')
+  //   if (assets.length == 0) {
+  //     Logger.log('All assets has webp file - aborting task', 'ImageConversionService')
+  //     return;
+  //   }
 
-        sharp(resizePath).resize(250).webp().toFile(webpPath, (err, info) => {
 
-          if (!err) {
-            this.assetRepository.update({ id: asset.id }, { webpPath: webpPath })
-          }
+  //   for (const asset of assets) {
+  //     const resizePath = asset.resizePath;
+  //     if (resizePath != '') {
+  //       const webpPath = resizePath.replace('jpeg', 'webp')
 
-        });
-      }
-    }
-  }
+  //       sharp(resizePath).resize(250).webp().toFile(webpPath, (err, info) => {
+
+  //         if (!err) {
+  //           this.assetRepository.update({ id: asset.id }, { webpPath: webpPath })
+  //         }
+
+  //       });
+  //     }
+  //   }
+  // }
 }
